[app][storage] Fix build issue due to ipc_msg_info structure updated.

+++ ipc.c
@@ -159,7 +159,7 @@ static int do_handle_msg(struct ipc_channel_context *ctx, const uevent_t *ev)
 	}
 
 	if (msg_inf.len > MSG_BUF_MAX_SIZE) {
-		TLOGE("%s: message too large %zd\n", __func__, msg_inf.len);
+		TLOGE("%s: message too large %d\n", __func__, msg_inf.len);
 		put_msg(chan, msg_inf.id);
 		return ERR_NOT_ENOUGH_BUFFER;
 	}
@@ -332,21 +332,21 @@ int sync_ipc_send_msg(handle_t session, iovec_t *tx_iovecs, size_t tx_iovec_coun
 		return rc;
 	}
 
-	size_t min_len = rx_iovecs[0].len;
+	uint32_t min_len = rx_iovecs[0].len;
 	if (inf.len < min_len) {
-		TLOGE("%s: invalid response length (%zd)\n", __func__, inf.len);
+		TLOGE("%s: invalid response length (%d)\n", __func__, inf.len);
 		put_msg(session, inf.id);
 		return ERR_NOT_VALID;
 	}
 
 	/* calculate total message size */
-	size_t resp_size = 0;
+	uint32_t resp_size = 0;
 	for (size_t i = 0; i < rx_iovec_count; ++i) {
 		resp_size += rx_iovecs[i].len;
 	}
 
 	if (resp_size < inf.len) {
-		TLOGE("%s: response buffer too short (%zd < %zd) \n", __func__,
+		TLOGE("%s: response buffer too short (%d < %d) \n", __func__,
 		      resp_size, inf.len);
 		put_msg(session, inf.id);
 		return ERR_BAD_LEN;
